requirements.txt
streamlit==1.37.1
pandas==2.2.2
numpy==2.0.1
plotly==5.23.0
scikit-learn==1.5.1
statsmodels==0.14.2
python-dateutil==2.9.0.post0
pillow==10.4.0
pytesseract==0.3.13
bcrypt==4.2.0
